payment-service/src/requirements.txt
flask
flask_sqlalchemy
flask_cors
python-dotenv
mysql-connector-python
requests
